+++ HtmlRun.Runtime/Constants/MathInstructionsSet.cs
@@ -3,4 +3,8 @@
 public static class MathInstructionsSet
 {
   public const string Clamp = "Clamp";
+
+  public const string IncrementValue = "IncrementValue";
+
+  public const string DecrementValue = "DecrementValue";
 }